#1 Split project into multiple modules
