+++ .gitignore
@@ -1,0 +1,4 @@
+analysis/out
+plugin/out
+simulation-client
+simulation-host
