+++ analysis/src/main/java/io/coupling/system/analysis/GraphObject.java
@@ -1,0 +1,8 @@
+package io.coupling.system.analysis;
+
+import org.neo4j.driver.v1.Session;
+
+interface GraphObject {
+
+  void persist(final Session session);
+}
