+++ build.gradle
@@ -1,18 +1,10 @@
 buildscript {
-    ext {
-        springBootVersion = '2.0.0.RELEASE'
-    }
     repositories {
         mavenCentral()
-    }
-    dependencies {
-        classpath("org.springframework.boot:spring-boot-gradle-plugin:${springBootVersion}")
     }
 }
 
 apply plugin: 'java'
-apply plugin: 'org.springframework.boot'
-apply plugin: 'io.spring.dependency-management'
 
 group 'io.coupling'
 version '1.0-SNAPSHOT'
@@ -21,30 +13,4 @@ sourceCompatibility = 1.8
 repositories {
     mavenCentral()
     maven { url "https://repo.spring.io/milestone" }
-}
-
-ext {
-    springCloudVersion = 'Finchley.M9'
-}
-
-dependencies {
-    compile 'org.springframework.boot:spring-boot-starter-web'
-    compile 'org.springframework.boot:spring-boot-starter-jdbc'
-    compile 'org.springframework.cloud:spring-cloud-starter-sleuth'
-    compile group: 'ch.qos.logback', name: 'logback-core', version: '1.3.0-alpha4'
-    compile group: 'net.logstash.logback', name: 'logstash-logback-encoder', version: '5.0'
-    compile group: 'com.h2database', name: 'h2', version: '1.4.197'
-    compile group: 'net.ttddyy', name: 'datasource-proxy', version: '1.4.7'
-    compile group: 'com.github.jsqlparser', name: 'jsqlparser', version: '1.1'
-    compile group: 'org.neo4j.driver', name: 'neo4j-java-driver', version: '1.4.4'
-    compile group: 'com.google.code.gson', name: 'gson', version: '2.8.2'
-    compile group: 'com.google.guava', name: 'guava', version: '24.1-jre'
-    testCompile group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.1.0'
-    testCompile group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.1.0'
-}
-
-dependencyManagement {
-    imports {
-        mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
-    }
 }
