--- src/main/java/io/coupling/systems/analysis/GraphObject.java
@@ -1,8 +1,0 @@
-package io.coupling.systems.analysis;
-
-import org.neo4j.driver.v1.Session;
-
-interface GraphObject {
-
-  void persist(final Session session);
-}
