--- src/main/resources/bootstrap.yml
@@ -1,1 +1,0 @@
-spring.application.name: simulation
