--- src/main/resources/logback.xml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-  <include resource="org/springframework/boot/logging/logback/defaults.xml"/>
-  ​
-  <springProperty scope="context" name="springAppName" source="spring.application.name"/>
-  <property name="LOG_FILE" value="${BUILD_FOLDER:-build}/${springAppName}"/>​
-
-  <property name="CONSOLE_LOG_PATTERN"
-    value="%clr(%d{yyyy-MM-dd HH:mm:ss.SSS}){faint} %clr(${LOG_LEVEL_PATTERN:-%5p}) %clr(${PID:- }){magenta} %clr(---){faint} %clr([%15.15t]){faint} %clr(%-40.40logger{39}){cyan} %clr(:){faint} %m%n${LOG_EXCEPTION_CONVERSION_WORD:-%wEx}"/>
-
-  <appender name="console" class="ch.qos.logback.core.ConsoleAppender">
-    <filter class="ch.qos.logback.classic.filter.ThresholdFilter">
-      <level>TRACE</level>
-    </filter>
-    <encoder>
-      <pattern>${CONSOLE_LOG_PATTERN}</pattern>
-      <charset>utf8</charset>
-    </encoder>
-  </appender>
-  ​
-  <appender name="logstash" class="ch.qos.logback.core.rolling.RollingFileAppender">
-    <file>${LOG_FILE}.json</file>
-    <rollingPolicy class="ch.qos.logback.core.rolling.TimeBasedRollingPolicy">
-      <fileNamePattern>${LOG_FILE}.json.%d{yyyy-MM-dd}.gz</fileNamePattern>
-      <maxHistory>7</maxHistory>
-    </rollingPolicy>
-    <encoder class="net.logstash.logback.encoder.LoggingEventCompositeJsonEncoder">
-      <providers>
-        <timestamp>
-          <timeZone>UTC</timeZone>
-        </timestamp>
-        <pattern>
-          <pattern>
-            {
-            "severity": "%level",
-            "service": "${springAppName:-}",
-            "trace": "%X{X-B3-TraceId:-}",
-            "span": "%X{X-B3-SpanId:-}",
-            "parent": "%X{X-B3-ParentSpanId:-}",
-            "exportable": "%X{X-Span-Export:-}",
-            "pid": "${PID:-}",
-            "thread": "%thread",
-            "class": "%logger{40}",
-            "rest": "%message"
-            }
-          </pattern>
-        </pattern>
-      </providers>
-    </encoder>
-  </appender>
-
-  <logger name="io.coupling.systems.simulation.RequestLoggingInterceptor" level="TRACE">
-    <appender-ref ref="logstash"/>
-  </logger>
-
-  <logger name="net.ttddyy.dsproxy.listener.logging.SLF4JQueryLoggingListener" level="TRACE">
-    <appender-ref ref="logstash"/>
-  </logger>
-  ​
-  <root level="INFO">
-    <appender-ref ref="console"/>
-  </root>
-</configuration>
