--- src/main/resources/schema.sql
@@ -1,1 +1,0 @@
-CREATE TABLE `information` (`id` BIGINT IDENTITY NOT NULL PRIMARY KEY, `data` VARCHAR(100) NOT NULL);
